services/vector_service/requirements.txt
# services/vector_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase-py[async]>=2.0.0   # For Supabase pgvector interaction (RPC, upsert)
google-generativeai>=0.4.0 # For Gemini embedding generation